@@ -14,12 +14,11 @@
     }
   },
   "require": {
-    "php": ">=7.0",
+    "php": ">=7.4",
     "ext-bcmath": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "^6.5",
-    "friendsofphp/php-cs-fixer": "2.16"
+    "phpunit/phpunit": "^6.5"
   },
   "config": {
     "sort-packages": true,
