@@ -8,13 +8,15 @@ use PHPUnit\Framework\TestCase;
 use Task1\CommissionTask\Service\Converter;
 
 class ConverterTest extends TestCase{
-    public function testConvertToEuro(){
+    public function testConvertToEuro() : void
+    {
         $converter = new Converter(1, 2, 3);
         $this->assertEquals(
             1, $converter->convertToEuros(3, 'JPY')
         );
     }
-    public function testConvertFromEuro(){
+    public function testConvertFromEuro() : void
+    {
         $converter = new Converter(1, 2, 3);
         $this->assertEquals(
             2, $converter->convertFromEuros(1, 'USD')
